+++ gtfs_kit/routes.py
@@ -761,17 +761,16 @@ def routes_to_geojson(
     If the Feed has no shapes, then raise a ValueError.
     If any of the given route IDs are not found in the feed, then raise a ValueError.
     """
-    if route_ids is not None:
-        D = set(route_ids) - set(feed.routes.route_id)
-        if D:
-            raise ValueError(f"Route IDs {D} not found in feed.")
+    if route_ids is None or not list(route_ids):
+        route_ids = feed.routes.route_id
+
+    D = set(route_ids) - set(feed.routes.route_id)
+    if D:
+        raise ValueError(f"Route IDs {D} not found in feed.")
 
     # Get routes
     g = geometrize_routes(feed, route_ids=route_ids, split_directions=split_directions)
-    if g.empty:
-        collection = {"type": "FeatureCollection", "features": []}
-    else:
-        collection = json.loads(g.to_json())
+    collection = json.loads(g.to_json())
 
     # Get stops if desired
     if include_stops: